--- src/const.py
@@ -1,3 +1,0 @@
-DATA_DIR = "/home/loaner/workspace/data"
-
-NUM_NEGATIVE = 10
