+++ src/data_queries.py
@@ -1,0 +1,240 @@
+import json
+import os
+import sys
+import time
+import numpy as np
+import pandas as pd
+from tqdm import tqdm
+import requests
+from sklearn.model_selection import train_test_split
+
+
+from util import DATA_DIR, NUM_NEGATIVE, DATASET_START_YEAR, DATASET_END_YEAR, CITATION_YEARS
+from util import KAGGLE_DATA_PATH, PAPERS_PATH, AUTHORS_PATH
+
+
+QUERY_FAIL_DELAY = 1
+QUERY_FAIL_MULT_DELAY = 2
+QUERY_DUMP_INTERVAL = 10
+
+
+def kaggle_json_to_parquet():
+    if os.path.exists(KAGGLE_DATA_PATH):
+        print("Kaggle data already converted to parquet - Skipping")
+        return
+    print("\nFiltering and converting Arxiv Kaggle data to Pandas")
+    with open(os.path.join(DATA_DIR, "arxiv-metadata-oai-snapshot.json")) as f:
+        kaggle_data = []
+        for l in tqdm(f.readlines(), "Parsing json"):
+            l = json.loads(l)
+            if "cs." in l["categories"]:
+                categories = l["categories"].split()
+                for c in categories:
+                    if c.startswith("cs"):
+                        l["categories"] = categories
+                        kaggle_data.append(l)
+                        break
+        kaggle_data = pd.DataFrame(kaggle_data)
+        print(f"Filtering relevant years (year < {DATASET_END_YEAR}) & (year >= {DATASET_START_YEAR})")
+        kaggle_data['update_date'] = pd.to_datetime(kaggle_data['update_date'])
+        kaggle_data['year_updated'] = kaggle_data['update_date'].dt.year
+        kaggle_data = kaggle_data[(kaggle_data["year_updated"] < DATASET_END_YEAR) & (kaggle_data["year_updated"] >= DATASET_START_YEAR)]
+        kaggle_data.to_parquet(KAGGLE_DATA_PATH)
+        print(kaggle_data)
+
+
+def get_citing_authors(l: list, paper_year: int, citation_years: int = CITATION_YEARS):
+    authors = set()
+    for citation in l:
+        try:
+            year = int(citation["year"])
+            if year < paper_year + citation_years:
+                for author in citation["authors"]:
+                    authors.add(int(author["authorId"]))
+        except TypeError:
+            # year is null
+            continue
+    return list(authors)
+
+
+def batch_query(
+    json_save_path,
+    query_ids,
+    batch_size,
+    query_fields,
+    query_url,
+    process_response_f
+):
+    if os.path.exists(json_save_path):
+        with open(json_save_path) as f:
+            data = json.load(f)
+    else:
+        data = {}
+    print(f"Number of query ids: {len(query_ids)}")
+    filtered_query_ids = [id for id in query_ids if str(id) not in data.keys()]
+    print(f"Ids without info: {len(filtered_query_ids)}")
+
+    fail_delay = 0
+    pbar = tqdm(total=len(query_ids))
+    pbar.update(len(query_ids) - len(filtered_query_ids))
+    idx = 0
+    prev_request_time = 0
+    while idx < len(filtered_query_ids):
+        batch_ids = filtered_query_ids[idx: idx + batch_size]
+        current_time = round(time.time() * 1000)  # time in milisecond
+        throughput_delay = prev_request_time + 1000 - current_time
+        throughput_delay = max(0, min(throughput_delay, 1000)) / 1000
+        if throughput_delay > 0:
+            time.sleep(throughput_delay)
+        prev_request_time = current_time
+        api_key = os.environ.get('API_KEY')
+        response = requests.post(
+            query_url,
+            headers={'x-api-key': api_key} if api_key is not None else {},
+            params={'fields': query_fields},
+            json={"ids": batch_ids}
+        )
+        if response.status_code != 200:
+            if "error" in response.text:
+                print(response.text)
+                sys.exit(1)
+            fail_delay = max(QUERY_FAIL_DELAY, fail_delay * QUERY_FAIL_MULT_DELAY)
+            print(f" - Sleeping for {fail_delay} seconds")
+            if "Too Many Requests" not in response.text:
+                print(json.loads(response.text))
+            time.sleep(fail_delay)
+            continue
+
+        for id, response in zip(batch_ids, response.json()):
+            if response is None:
+                print(f"{id} returned None")
+                data[id] = None
+                continue
+            data[id] = process_response_f(response)
+        
+        idx += batch_size
+        pbar.update(batch_size)
+        fail_delay = 0
+
+        if idx % (batch_size * QUERY_DUMP_INTERVAL) == 0:
+            # print(f"Dumping data to {json_save_path}")
+            with open(json_save_path, 'w') as f:
+                json.dump(data, f, indent=4)
+    pbar.close()
+
+    with open(json_save_path, 'w') as f:
+        json.dump(data, f, indent=4)
+
+
+def prepare_papers_data():
+    print("\nQuering Semantic Scholar for papers info")
+    kaggle_data = pd.read_parquet(KAGGLE_DATA_PATH)
+
+    # Work on few papers
+    kaggle_data = kaggle_data.sample(frac=1., random_state=0)
+    kaggle_data = kaggle_data[:100]
+    
+    def process_paper_response(j: json):
+        if j["year"] is None:
+            return None
+        j["authors"] = list(set([int(tmp["authorId"]) for tmp in j["authors"] if tmp["authorId"] is not None]))
+        j["citing_authors"] = get_citing_authors(j["citations"], int(j["year"]))
+        del j["citations"]
+        j["cited_authors"] = list(set([int(author["authorId"]) for ref in j["references"] for author in ref["authors"] if author["authorId"] is not None]))
+        del j["references"]
+        del j["paperId"]
+        return j
+
+    # See Semantic Scholar API docs: https://api.semanticscholar.org/api-docs/graph#tag/Paper-Data/operation/post_graph_get_papers
+    batch_query(
+        json_save_path=PAPERS_PATH,
+        query_ids=[f"ARXIV:{id}" for id in kaggle_data["id"]],
+        batch_size=20,
+        query_fields="year,referenceCount,isOpenAccess,publicationTypes,authors,citations.year,citations.authors,references.authors",
+        query_url="https://api.semanticscholar.org/graph/v1/paper/batch",
+        process_response_f=process_paper_response
+    )
+
+
+def prepare_authors_data():
+    print("\nQuering Semantic Scholar for authors info")
+    with open(PAPERS_PATH) as f:
+        papers = json.load(f)
+    citing_authors = set()
+    for paper in papers.values():
+        for citing_author in paper["citing_authors"]:
+            citing_authors.add(citing_author)
+    print(f"{len(papers)} papers have {len(citing_authors)} citing authors")
+
+    def process_author_response(j: json):
+        papers = [
+            {
+                "year": paper["year"],
+                "fieldsOfStudy": paper["fieldsOfStudy"],
+                "s2FieldsOfStudy": list(set([tmp["category"] for tmp in paper["s2FieldsOfStudy"] if tmp["category"] is not None]))
+            }
+            for paper in j["papers"] if paper["year"] is not None
+        ]
+        return {
+            "papers": papers
+        }
+
+    # See Semantic Scholar API docs: https://api.semanticscholar.org/api-docs/graph#tag/Author-Data/operation/post_graph_get_authors
+    batch_query(
+        json_save_path=AUTHORS_PATH,
+        query_ids=list(citing_authors),
+        batch_size=100,
+        query_fields="papers.year,papers.fieldsOfStudy,papers.s2FieldsOfStudy",
+        query_url="https://api.semanticscholar.org/graph/v1/author/batch",
+        process_response_f=process_author_response
+    )
+
+
+def generate_samples():
+    print("\nGenerating train, validation and test folds")
+    rng = np.random.default_rng(seed=42)
+    with open(PAPERS_PATH) as f:
+        papers = json.load(f)
+    with open(AUTHORS_PATH) as f:
+        authors = json.load(f)
+    author_keys = list(authors.keys())
+
+    samples = []
+    for paper_id, paper in papers.items():
+        # Positive samples
+        for citing_author in paper["citing_authors"]:
+            samples.append(
+                {
+                    "paper": paper_id,
+                    "year": paper["year"],
+                    "author": citing_author,
+                    "label": True
+                }
+            )
+        # Negative samples
+        for author_idx in rng.integers(low=0, high=len(authors), size=NUM_NEGATIVE):
+            samples.append(
+                {
+                    "paper": paper_id,
+                    "year": paper["year"],
+                    "author": author_keys[author_idx],
+                    "label": False
+                }
+            )
+
+    samples = pd.DataFrame.from_records(samples)
+    test = samples[samples["year"] == 2020]
+    train_validation_samples = samples[samples["year"] < 2020]
+    train, validation = train_test_split(train_validation_samples, test_size=0.2, stratify=train_validation_samples[["year", "label"]], random_state=42)
+    for d, name in [(train, "train"), (validation, "validation"), (test, "test")]:
+        print(f"{name}:", len(d))
+        d = d.drop("year", axis=1)
+        d.to_csv(os.path.join(DATA_DIR, f"{name}.csv"), index=False)
+
+
+if __name__ == '__main__':
+    kaggle_json_to_parquet()
+    prepare_papers_data()
+    prepare_authors_data()
+    generate_samples()
+
