+++ src/models/base_model.py
@@ -1,0 +1,3 @@
+class base_model:
+    def save():
+        raise NotImplementedError("Save for base_model should not be called")
