+++ src/train.py
@@ -1,0 +1,11 @@
+import json
+
+from data import Data
+
+
+def train():
+    pass
+
+
+if __name__ == '__main__':
+    data = Data()
